requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
bcrypt
pyjwt
pytest
python-multipart
python-dotenv